+++ NEWS.md
@@ -1,8 +1,6 @@
 # inlabru (development version)
 
-* Remove inactive/unused `maptools` dependency #178
-
-* Remove last active `rgdal` use, to prepare for complete rgdal dependency removal #178
+* Remove `rgdal` and `maptools` dependencies #178
 
 # inlabru 2.7.0